ocr_pdf_annotator/requirements.txt
fastapi
uvicorn
pillow
pdf2image
pytesseract
